backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
pandas
numpy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pymysql
alembic
apscheduler
scikit-learn
bcrypt==3.2.2
